server/requirements.txt
flask==2.0.1
flask-cors==3.0.10
SQLAlchemy==2.0.27
python-dotenv==1.0.1
alembic==1.13.1
werkzeug==2.0.3
jieba==0.42.1 
alembic
#chromadb
#sentence-transformers
